+++ ft_memccpy.c
@@ -6,15 +6,15 @@
 /*   By: aleon-ca <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/06 12:54:23 by aleon-ca          #+#    #+#             */
-/*   Updated: 2019/11/18 14:23:04 by aleon-ca         ###   ########.fr       */
+/*   Updated: 2019/11/21 14:57:57 by aleon-ca         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memccpy(void *dest, const void *src, int c, unsigned int n)
+void	*ft_memccpy(void *dest, const void *src, int c, size_t n)
 {
-	unsigned int i;
+	size_t	i;
 
 	i = -1;
 	while (++i < n)
