+++ ft_memchr.c
@@ -6,15 +6,15 @@
 /*   By: aleon-ca <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/06 16:19:10 by aleon-ca          #+#    #+#             */
-/*   Updated: 2019/11/18 14:23:13 by aleon-ca         ###   ########.fr       */
+/*   Updated: 2019/11/21 14:59:23 by aleon-ca         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memchr(const void *str, int c, unsigned int n)
+void	*ft_memchr(const void *str, int c, size_t n)
 {
-	unsigned int i;
+	size_t i;
 
 	i = 0;
 	while (i < n)
