+++ ft_memcpy.c
@@ -6,15 +6,15 @@
 /*   By: aleon-ca <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/06 12:37:19 by aleon-ca          #+#    #+#             */
-/*   Updated: 2019/11/18 14:36:45 by aleon-ca         ###   ########.fr       */
+/*   Updated: 2019/11/21 14:57:10 by aleon-ca         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memcpy(void *dest, const void *src, unsigned int n)
+void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	unsigned int i;
+	size_t i;
 
 	if (n != 0 && (dest || src))
 	{
