+++ ft_memmove.c
@@ -6,15 +6,15 @@
 /*   By: aleon-ca <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/06 16:03:51 by aleon-ca          #+#    #+#             */
-/*   Updated: 2019/11/20 09:34:42 by aleon-ca         ###   ########.fr       */
+/*   Updated: 2019/11/21 14:58:45 by aleon-ca         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memmove(void *dest, const void *src, unsigned int len)
+void	*ft_memmove(void *dest, const void *src, size_t len)
 {
-	unsigned int	i;
+	size_t i;
 
 	if (len != 0 && (dest || src))
 	{
